requirements.txt
streamlit~=1.18.0
psutil~=5.9.0
pyttsx3~=2.90
speechrecognition~=3.8.1
wikipedia-api~=0.5.4
pyautogui~=0.9.53
pyjokes~=0.6
